+++ src/minishell1.c
@@ -74,4 +74,3 @@ int	main(int argc, char **argv, char **env)
 	}
 	return (0);
 }
-